# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for Osau AMS.",
        "title": "Osau AMS API",
        "contact": {
            "name": "API Support",
            "email": "bakit.german.work@icloud.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/admins/universities": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a list of all universities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Universities"
                ],
                "summary": "Get all universities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.University"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing university",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Universities"
                ],
                "summary": "Update a university",
                "parameters": [
                    {
                        "description": "University info",
                        "name": "university",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PutUniversityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new university",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Universities"
                ],
                "summary": "Create a university",
                "parameters": [
                    {
                        "description": "University info",
                        "name": "university",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateUniversityRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Partially update an existing university",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Universities"
                ],
                "summary": "Partially update a university",
                "parameters": [
                    {
                        "description": "University info",
                        "name": "university",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PatchUniversityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admins/universities/name/{name}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a university by its name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Universities"
                ],
                "summary": "Get a university by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "University Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.University"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admins/universities/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a university by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Universities"
                ],
                "summary": "Get a university by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "University ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.University"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete an existing university",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Universities"
                ],
                "summary": "Delete a university",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "University ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.University": {
            "type": "object",
            "properties": {
                "head_first_name": {
                    "type": "string"
                },
                "head_last_name": {
                    "type": "string"
                },
                "head_middle_name": {
                    "type": "string"
                },
                "university_email": {
                    "type": "string"
                },
                "university_id": {
                    "type": "integer"
                },
                "university_name": {
                    "type": "string"
                }
            }
        },
...
        "handler.CreateUniversityRequest": {
            "type": "object",
            "required": [
                "head_first_name",
                "head_last_name",
                "head_middle_name",
                "university_email",
                "university_name"
            ],
            "properties": {
                "head_first_name": {
                    "type": "string",
                    "maxLength": 60,
                    "minLength": 2
                },
                "head_last_name": {
                    "type": "string",
                    "maxLength": 60,
                    "minLength": 2
                },
                "head_middle_name": {
                    "type": "string",
                    "maxLength": 60,
                    "minLength": 2
                },
                "university_email": {
                    "type": "string",
                    "maxLength": 60,
                    "minLength": 8
                },
                "university_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.PatchUniversityRequest": {
            "type": "object",
            "required": [
                "university_id"
            ],
            "properties": {
                "head_first_name": {
                    "type": "string",
                    "maxLength": 60,
                    "minLength": 2
                },
                "head_last_name": {
                    "type": "string",
                    "maxLength": 60,
                    "minLength": 2
                },
                "head_middle_name": {
                    "type": "string",
                    "maxLength": 60,
                    "minLength": 2
                },
                "university_email": {
                    "type": "string",
                    "maxLength": 60,
                    "minLength": 8
                },
                "university_id": {
                    "type": "integer"
                },
                "university_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
...
        "handler.PutUniversityRequest": {
            "type": "object",
            "required": [
                "head_first_name",
                "head_last_name",
                "head_middle_name",
                "university_email",
                "university_id",
                "university_name"
            ],
            "properties": {
                "head_first_name": {
                    "type": "string",
                    "maxLength": 60,
                    "minLength": 2
                },
                "head_last_name": {
                    "type": "string",
                    "maxLength": 60,
                    "minLength": 2
                },
                "head_middle_name": {
                    "type": "string",
                    "maxLength": 60,
                    "minLength": 2
                },
                "university_email": {
                    "type": "string",
                    "maxLength": 60,
                    "minLength": 8
                },
                "university_id": {
                    "type": "integer"
                },
                "university_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
...
        "handler.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}